[system-loaders] Siege::System::Loader has abstract semantic - Loader::Abstract is useless
